move contract bindings into sub folder
