+++ src/lib.rs
@@ -1,8 +1,7 @@
-pub mod conduit_controller;
+pub mod bindings;
 pub mod configuration;
 pub mod indexer;
 pub mod routes;
-pub mod seaport;
 pub mod startup;
 pub mod structs;
 pub mod telemetry;
